# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server server.",
        "title": "Fiber Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/user/follow/:followId": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add followId user to your following list",
                "tags": [
                    "User-Follow"
                ],
                "summary": "Follow User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id on the user want to follow",
                        "name": "followId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseOKModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/user/followers/:userId/:skip/:limit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get user followers",
                "tags": [
                    "User-Follow"
                ],
                "summary": "Followers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "skip",
                        "name": "skip",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FollowUserDataModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/user/followings/:userId/:skip/:limit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get user followings",
                "tags": [
                    "User-Follow"
                ],
                "summary": "Followings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "skip",
                        "name": "skip",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FollowUserDataModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/user/getToken": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get new Tokens, also return `refreshToken`",
                "tags": [
                    "User-Auth"
                ],
                "summary": "Get Token",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "return refreshToken in response body instead of saving in cookie",
                        "name": "noCookie",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "also return profile images, slower response",
                        "name": "profileImages",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Device Info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeviceInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserViewModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/user/login": {
            "post": {
                "description": "Login with provided credentials",
                "tags": [
                    "User-Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "return refreshToken in response body instead of saving in cookie",
                        "name": "noCookie",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "User object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginViewModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserViewModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/user/logout": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Logout user, return accessToken as empty string and also reset/remove refreshToken cookie if use in browser\n.in other environments reset refreshToken from client after successful logout.",
                "tags": [
                    "User-Auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
...
        "/v1/user/setNotifToken": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "send device token as Notification token",
                "tags": [
                    "User"
                ],
                "summary": "Notification Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "notifToken",
                        "name": "notifToken",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/user/signup": {
            "post": {
                "description": "Register a new user with the provided credentials\nDevice detection can be improved on client side with adding 'deviceInfo.fingerprint'",
                "tags": [
                    "User-Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "return refreshToken in response body instead of saving in cookie",
                        "name": "noCookie",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "User object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterViewModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserViewModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/user/unfollow/:followId": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove followId user from users following list",
                "tags": [
                    "User-Follow"
                ],
                "summary": "UnFollow User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id on the user want to unfollow",
                        "name": "followId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseOKModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.DeviceInfo": {
            "type": "object",
            "properties": {
                "appName": {
                    "type": "string"
                },
                "appVersion": {
                    "type": "string"
                },
                "deviceModel": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "notifToken": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                }
            }
        },
        "model.FollowListProfileImageDataModel": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.FollowUserDataModel": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "profileImages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FollowListProfileImageDataModel"
                    }
                },
                "publicName": {
                    "type": "string"
                },
                "rawUsername": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.LoginViewModel": {
            "type": "object",
            "properties": {
                "deviceInfo": {
                    "$ref": "#/definitions/model.DeviceInfo"
                },
                "password": {
                    "type": "string"
                },
                "username_email": {
                    "type": "string"
                }
            }
        },
...
        "model.ProfileImage": {
            "type": "object",
            "properties": {
                "addDate": {
                    "type": "string"
                },
                "originalSize": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "model.RegisterViewModel": {
            "type": "object",
            "properties": {
                "confirmPassword": {
                    "type": "string"
                },
                "deviceInfo": {
                    "$ref": "#/definitions/model.DeviceInfo"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.TokenViewModel": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "accessToken_expire": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
...
        "model.UserViewModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "profileImages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProfileImage"
                    }
                },
                "token": {
                    "$ref": "#/definitions/model.TokenViewModel"
                },
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "response.ResponseErrorModel": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errorMessage": {}
            }
        },
        "response.ResponseOKModel": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errorMessage": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}